fix(): change paymentAmount from Double to Int -> stripe needs Int that represent cents

+++ src/main/kotlin/ecommerce/dto/OrderDto.kt
@@ -5,7 +5,7 @@ class OrderDto(
     val createdAt: LocalDateTime,
     var status: String,
     val items: List<OrderItemDto>,
-    val paymentAmount: Double,
+    val paymentAmount: Int,
     val currency: String,
     val paymentMethod: String,
     var checkoutSessionId: String? = null,
